+++ lib/views/course_view/list_courses_view/my_courses_view.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:gocast_mobile/views/course_view/components/course_card_view.dart';
 
-import '../utils/constants.dart';
-import 'components/course_screen.dart';
+import '../../utils/constants.dart';
+import '../components/course_screen.dart';
 
 // MyCourses.dart
 class MyCourses extends StatelessWidget {
